+++ src/Rector/Misc/BuildTaskUpdateRector.php
@@ -4,15 +4,26 @@ declare(strict_types=1);
 
 namespace Netwerkstatt\SilverstripeRector\Rector\Misc;
 
+use Netwerkstatt\SilverstripeRector\Rector\Misc\BuildTaskUpdateRector\RequestToInputOptionVisitor;
 use PhpParser\Node;
+use PhpParser\Node\Arg;
+use PhpParser\Node\Expr\Array_;
+use PhpParser\Node\Expr\ArrayItem;
+use PhpParser\Node\Expr\ClassConstFetch;
+use PhpParser\Node\Expr\MethodCall;
+use PhpParser\Node\Expr\New_;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Identifier;
 use PhpParser\Node\Name\FullyQualified;
 use PhpParser\Node\Param;
+use PhpParser\Node\Scalar\String_;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassMethod;
+use PhpParser\Node\Stmt\Expression;
 use PhpParser\Node\Stmt\Property;
+use PhpParser\Node\Stmt\Return_;
 use PhpParser\Node\VarLikeIdentifier;
+use PhpParser\NodeTraverser;
 use PHPStan\PhpDocParser\Ast\PhpDoc\GenericTagValueNode;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTagNode;
 use PHPStan\Type\ObjectType;
@@ -104,8 +115,22 @@ CODE_SAMPLE
         $hasChanged = false;
 
         foreach ($node->getProperties() as $property) {
+            if ($this->isName($property, 'segment')) {
+                $property->props[0]->name = new VarLikeIdentifier('commandName');
+                $property->type = new Identifier('string');
+                $property->flags = Class_::MODIFIER_PROTECTED | Class_::MODIFIER_STATIC;
+                $hasChanged = true;
+            }
+
             if ($this->isName($property, 'title')) {
-                $property->props[0]->name = new Node\VarLikeIdentifier('description');
+                $property->type = new Identifier('string');
+                $property->flags = Class_::MODIFIER_PROTECTED;
+                $hasChanged = true;
+            }
+
+            if ($this->isName($property, 'description')) {
+                $property->type = new Identifier('string');
+                $property->flags = Class_::MODIFIER_PROTECTED | Class_::MODIFIER_STATIC;
                 $hasChanged = true;
             }
         }
@@ -113,19 +138,35 @@ CODE_SAMPLE
         $runMethod = $node->getMethod('run');
         if ($runMethod instanceof ClassMethod) {
             $runMethod->name = new Identifier('execute');
-            
+            $runMethod->flags = Class_::MODIFIER_PROTECTED;
+            $runMethod->returnType = new Identifier('int');
+
+            $inputVarName = 'input';
+            $outputVarName = 'output';
+
             $runMethod->params = [
                 new Param(
-                    new Variable('input'),
+                    new Variable($inputVarName),
                     null,
                     new FullyQualified('Symfony\Component\Console\Input\InputInterface')
                 ),
                 new Param(
-                    new Variable('output'),
+                    new Variable($outputVarName),
                     null,
-                    new FullyQualified('Symfony\Component\Console\Output\OutputInterface')
+                    new FullyQualified('SilverStripe\PolyExecution\PolyOutput')
                 ),
             ];
+
+            $visitor = new RequestToInputOptionVisitor($inputVarName, $outputVarName);
+            $traverser = new NodeTraverser();
+            $traverser->addVisitor($visitor);
+            /** @var Node\Stmt[] $stmts */
+            $stmts = (array) $traverser->traverse((array) $runMethod->stmts);
+            $runMethod->stmts = $stmts;
+
+            $runMethod->stmts[] = new Return_(
+                $this->nodeFactory->createClassConstFetch('Symfony\Component\Console\Command\Command', 'SUCCESS')
+            );
 
             $phpDocInfo = $this->phpDocInfoFactory->createFromNodeOrEmpty($runMethod);
             $phpDocInfo->addPhpDocTagNode(new PhpDocTagNode(
@@ -137,9 +178,43 @@ CODE_SAMPLE
                 new GenericTagValueNode('Define input parameters in getOptions() if needed.')
             ));
 
+            $options = $visitor->getOptions();
+            if ($options !== []) {
+                $node->stmts[] = $this->createGetOptionsMethod($options);
+            }
+
             $hasChanged = true;
         }
 
         return $hasChanged ? $node : null;
+    }
+
+    private function createGetOptionsMethod(array $options): ClassMethod
+    {
+        $inputOptionClass = 'Symfony\Component\Console\Input\InputOption';
+        $optionNodes = [];
+
+        foreach ($options as $option) {
+            $optionNodes[] = new ArrayItem(
+                new New_(
+                    new FullyQualified($inputOptionClass),
+                    [
+                        new Arg(new String_($option)),
+                        new Arg($this->nodeFactory->createNull()),
+                        new Arg($this->nodeFactory->createClassConstFetch($inputOptionClass, 'VALUE_NONE')),
+                        new Arg(new String_('do something specific')), // Default description as in issue
+                    ]
+                )
+            );
+        }
+
+        $method = new ClassMethod('getOptions');
+        $method->flags = Class_::MODIFIER_PUBLIC;
+        $method->returnType = new Identifier('array');
+        $method->stmts = [
+            new Return_(new Array_($optionNodes))
+        ];
+
+        return $method;
     }
 }